apps/api/llm-integration/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Database & caching
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.1.18
redis==5.0.8

# Parsing & NLP
pdfplumber==0.11.4
python-docx==1.1.2
python-magic-bin==0.4.14; sys_platform == 'win32'
python-magic==0.4.27; sys_platform != 'win32'
spacy==3.7.5
dateparser==1.2.0
phonenumbers==8.13.45
email-validator==2.2.0
tldextract==5.1.2
rapidfuzz==3.9.7

# LLM Integration
openai==1.54.5
anthropic==0.39.0

# Security & hashing
argon2-cffi==23.1.0
python-multipart==0.0.9

# Observability
structlog==24.4.0
python-json-logger==2.0.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
hypothesis==6.112.1
httpx==0.27.2

# Code quality
ruff==0.6.8
black==24.8.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
